# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_pos-retail_go_backend_internal_types_request.CreateSaleRequest:
    properties:
      customer_id:
        type: string
      items:
        items:
          $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_request.SaleItemRequest'
        type: array
      loyalty_points_redeem:
        type: integer
      notes:
        type: string
      payments:
        items:
          $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_request.SalePaymentRequest'
        type: array
      warehouse_id:
        type: string
    type: object
...
  github_com_pos-retail_go_backend_internal_types_request.SaleItemRequest:
    properties:
      product_id:
        type: string
      promotion_code:
        type: string
      quantity:
        type: integer
    type: object
  github_com_pos-retail_go_backend_internal_types_request.SalePaymentRequest:
    properties:
      amount:
        type: number
      card_last_4:
        type: string
      method:
        type: string
      reference_number:
        type: string
    type: object
...
  github_com_pos-retail_go_backend_internal_types_response.ApiResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
  github_com_pos-retail_go_backend_internal_types_response.PaginatedResponse:
    properties:
      data: {}
      pagination:
        $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.Pagination'
      success:
        type: boolean
    type: object
  github_com_pos-retail_go_backend_internal_types_response.Pagination:
    properties:
      has_more:
        type: boolean
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
info:
  contact: {}
  description: Go (Fiber) implementation of POS Retail Backend.
  title: POS Retail Backend API
  version: "1.0"
paths:
...
  /api/sales:
    get:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Warehouse ID
        in: query
        name: warehouse_id
        type: string
      - description: Customer ID
        in: query
        name: customer_id
        type: string
      - description: Cashier ID
        in: query
        name: cashier_id
        type: string
      - description: Sale status
        in: query
        name: status
        type: string
      - description: From date (YYYY-MM-DD)
        in: query
        name: date_from
        type: string
      - description: To date (YYYY-MM-DD)
        in: query
        name: date_to
        type: string
      - description: Sale number search
        in: query
        name: sale_number
        type: string
      - default: 50
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.PaginatedResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
      security:
      - BearerAuth: []
      summary: List sales
      tags:
      - Sales
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Sale payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_request.CreateSaleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Create sale
      tags:
      - Sales
  /api/sales/{id}:
    get:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Sale ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Get sale
      tags:
      - Sales
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"